Fix signals,

+++ process.go
@@ -8,35 +8,34 @@ import (
 	"syscall"
 )
 
-type Process struct {
+type process struct {
 	*exec.Cmd
 	command []string
 }
 
-func NewProcess(command []string) *Process {
-	return &Process{command: command}
+func NewProcess(command []string) *process {
+	return &process{command: command}
 }
 
-func (p *Process) Start() (err error) {
+func (p *process) Start() error {
 	p.Cmd = exec.Command(p.command[0], p.command[1:]...)
 	p.Cmd.Stdout = os.Stdout
 	p.Cmd.Stderr = os.Stderr
 	p.Cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 
-	err = p.Cmd.Start()
-	return
+	return p.Cmd.Start()
 }
 
-func (p *Process) Stop() {
-	if p.Process != nil {
+func (p *process) Stop() {
+	if p.Cmd != nil && p.Cmd.Process != nil {
 		// TODO: this works only in *nix world
 		syscall.Kill(-p.Process.Pid, syscall.SIGKILL)
 	}
 }
 
-func (p *Process) String() string {
+func (p *process) String() string {
 	s := fmt.Sprintf("`%s`", strings.Join(p.command, " "))
-	if p.Process != nil {
+	if p.Cmd != nil && p.Cmd.Process != nil {
 		s += fmt.Sprintf(" (pid: %d)", p.Process.Pid)
 	}
